+++ src/NetStandard/E.Standard.WebGIS.Tools/MapMarkup/Export/GpxExport.cs
@@ -41,7 +41,10 @@ class GpxExport : IExport
                 string toolType = feature.GetPropery<string>("_meta.tool");
                 string text = feature.GetPropery<string>("_meta.text");
 
-                if (toolType == "line")
+                if (toolType == "line"
+                    || (feature.Geometry?.type.Equals("linestring", StringComparison.OrdinalIgnoreCase) == true ||
+                        feature.Geometry?.type.Equals("multilinestring", StringComparison.OrdinalIgnoreCase) == true)
+                        )
                 {
                     var trk = GpxHelper.FromPolyline(_gpx, feature.ToShape() as Polyline);
                     if (trk != null)
@@ -51,7 +54,8 @@ class GpxExport : IExport
                         _featuresCount++;
                     }
                 }
-                else if (toolType == "symbol" || toolType == "text" || toolType == "point")
+                else if ((toolType == "symbol" || toolType == "text" || toolType == "point")
+                         || (feature.Geometry?.type.Equals("point", StringComparison.OrdinalIgnoreCase) == true))
                 {
                     wptType pointType = GpxHelper.FromPoint(feature.ToShape() as Point);
                     if (pointType != null)