+++ changelog.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Added
 ## Fixed
 
+- Download GPX from measure/edit sketch not worked correctly
+
 ## 8.26.901
 
 ## Added